almost decent prototype
fully recreated person class from the last few saves, now its pretty good for the most part. the only thing left to do is code the people to stay in their proper lanes when going/leaving a store. When a person leaves a store, +$10 is added to the store's revenue.

Price scout main: recreated numPeople for loop. Now they have a randomized amount of money. I also created a little personColor line of code for later, when we organize people's colours based on which stores they are moving in and out from. I also make the solid line form the main road into a dotted line because it doesnt make sense for people to cross over a solid line.